refactor(ui): extract layout-placement helpers from store/layout.ts
Splits the pure placement types/helpers (SurfaceId, WorkspaceLayout,
placeInLayout, removeSurface, repositionInLayout, isSurfaceFloor, etc.)
into store/layout-placement.ts, re-exported unchanged from layout.ts so
none of the 14 importing files move. Keeps layout.ts under the 300-line
cap ahead of the url-tab tasks landing more code in it (#281 task 32).
